clean architecture structure
