+++ src/main/java/com/fiap/hackaton/application/gateways/ProcessFileUploadSpec.java
@@ -1,0 +1,5 @@
+package com.fiap.hackaton.application.gateways;
+
+public interface ProcessFileUploadSpec {
+    void execute(String key);
+}
