+++ src/main/java/com/fiap/hackaton/domain/enums/UploadStatus.java
@@ -1,0 +1,7 @@
+package com.fiap.hackaton.domain.enums;
+
+public enum UploadStatus {
+    PENDING,
+    PROCESSED,
+    NOTIFIED
+}
