+++ src/main/java/com/fiap/hackaton/domain/exceptions/NotFoundException.java
@@ -1,0 +1,12 @@
+package com.fiap.hackaton.domain.exceptions;
+
+public class NotFoundException extends RuntimeException {
+
+    public NotFoundException() {
+        super();
+    }
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
