+++ src/main/java/com/fiap/hackaton/infrastructure/services/RemoveAsyncFileProcessedEventService.java
@@ -1,6 +1,6 @@
-package com.fiap.hackaton.service;
+package com.fiap.hackaton.infrastructure.services;
 
-import com.fiap.hackaton.broker.dto.FileProcessedEvent;
+import com.fiap.hackaton.infrastructure.presentation.workers.dto.FileProcessedEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
@@ -12,7 +12,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class FileProcessedEventService {
+public class RemoveAsyncFileProcessedEventService {
     private final ApplicationEventPublisher publisher;
 
     public void emitEvent(List<Path> filesTemp) {
